archsigのAIR系モジュール境界を分離する
